requirements.txt
# Minimal requirements for AIPL Chatbot
streamlit>=1.28.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
openai>=1.6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10
faiss-cpu>=1.7.0
rank-bm25>=0.2.2
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0
pypdf2>=3.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.0.0
matplotlib>=3.5.0
